Implementing the TCL example
I am trying to implement the TCL example of Insoon Yang's paper. To
this end, I had to modify the functions generateNoise and generatePartition
so that they are able to generate the noise and partition for the TCL example.

I have also added a class VectorFieldTCL that implements the dynamics of Insoon's
paper. And I am currently creating a function TCL that will mimic Fishery_management,
but now to the new vector field.

The class StatePartition had to be modified as we are now able to partition 2 and 3D spaces.
All the methods of this class have been modified.

I have created a function NotImplemented in case a vector field that is not implemented
is passed as arguments for the generatePartition and generateNoise functions.

The function MainValueFunctionIteartion.m was part of the Fishery_management file is know
and explicit function of Matlab. This is important to allow code reuse. In making these changes
I have also modified the function to deal with the TCL vector field.

Similar comments can be made with the EstimateTransition function.

The class ComputeValueFunction has also undergone the necessary changes to deal with
the TCL vector field.

This version of the code runs Fishery_management.m with no problem.