The functions ComputeValueFunction, EstimateTransition, MainValueFunctionsIteration, generatePartition and getDateSaveFiel have all been modified to permit the implementation of the different vector field. Most of the changes that I have made was to add some switch condition in some of the functions and methods in these files. For instance, the StateParti- tion class can now deal with one-dimensional spaces rather than 3-dimensional ones.
I have also found a bug in the way I was creating objects of the class MomentBasedAmbiguity.
Possibly, there is a bug in the Verify arguments functions. I decided to simplify the creation
of these objects by assigning an order in which each argument is present.

I have also created a function PlotResults that is able to generate the plots based on the data
that is saved under the ./results directory.

The current version of the code is able to run all types of Ambiguities for both vector fields, namely
'TCL' and 'Fishery'.