In this version of the code I have stopped using the HashTable class function because it was too inefficient. To this end, I have created a function createXandUString.m to unify the list and use the built-in containers.Map class of matlab.
I have also chenged the implementation of the KernelBasedAmbiguity.m class as it was also
inefficient. The kernel implementation is now in a separate function.

This version of the code works well with the Fhishery vector field. MomentBasedAmbiguity
class sometimes leads to unfeasible problem. Need to check this.

I have not tested this version of the code with the TCL vector field.