This commit fixes the counting problem that was appearing in previous versions of this code when updating the wait bar. This was due to the fact that the TimeHorizon parameter was an int8 variable.
I have also realised that the safety computation was wrong in the previous
version of the code. This has been mixed up when I was migrating from my
HashTable class to containers.Map. The issue was that we need to have ListX
rather than List in iterateValueFunction method in the class ComputeValueFunction.

Another feature that I have added in this version is to pass structures that contain
information about the ambiguity sets externally, as you can see in the relevant .m files.

I have also changed the way we create MomentBasedAmbuigty objects.

Another minor change was the modification on the partition for the TCL vector field.