In this commit I have changed the KernelAmbiguity class to include the term ||v -c||, where v is the value function at the next iteration and c is the estimation of the transtition kernel.
I have modified the plot function to be consistent with the way I am
saving the data in the other routines, especially due to the fact that I
have to add a montecarlo simulation funtion (as you can see in
MonteCarloSimulation.m file) to test the designed policy.

In the current version of the code, once the data is generated by
calling testTCL (or testFishery), we just need to get the variable in
the .mat file called path and pass it on to the MonteCarloSimulation
function, and then to PlotAllResults.